move api connection error handlers to node-core
