+++ packages/node-core/src/api.service.ts
@@ -1,8 +1,9 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
+import {IApiConnectionSpecific} from './api.connection';
 import {NetworkMetadataPayload} from './events';
-import {ApiConnectionError, ApiErrorType, ConnectionPoolService, ISubqueryProject} from './indexer';
+import {ApiConnectionError, ApiErrorType, ConnectionPoolService} from './indexer';
 import {getLogger} from './logger';
 
 const logger = getLogger('api');
@@ -14,12 +15,6 @@ export interface IApi<A = any, SA = any, B = any> {
   safeApi(height: number): SA;
   unsafeApi: A;
   networkMeta: NetworkMetadataPayload;
-}
-
-export interface IApiConnectionSpecific<A = any, SA = any, B = any> extends IApi<A, SA, B> {
-  handleError(error: Error): ApiConnectionError;
-  apiConnect(): Promise<void>;
-  apiDisconnect(): Promise<void>;
 }
 
 export abstract class ApiService<A = any, SA = any, B = any> implements IApi<A, SA, B> {
